requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
requests==2.26.0
factory-boy==3.2.0
Faker==8.16.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.3.0
pyyaml==5.4.1
uritemplate==3.0.1
coreapi==2.3.3
redis==3.5.3
celery==5.1.2
django-celery-beat==2.2.1
coverage
